ai-model-validation-platform/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client & CORS
httpx==0.25.2
fastapi-cors==0.0.6

# File Handling
python-magic==0.4.27
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# WebSocket & Real-time Communication
python-socketio==5.11.0
python-engineio>=4.8.0

# AI/ML dependencies - Full ML stack for complete functionality
torch>=2.0.0
torchvision>=0.15.0
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.25.2
scipy==1.11.4
matplotlib==3.8.2
pandas==2.1.4

# PostgreSQL Driver
psycopg2-binary

# Signal Acquisition Hardware (Optional - system will work without it)
# labjack-ljm==1.21.0  # LabJack voltage signal acquisition - install manually if needed
